Fix initialization if ocean::Cavity::cbasins.
IceModelVec2S::read() assumes that the file we're reading from has the
grid that matches the current grid PISM is running on. The "basins"
field usually comes from a file on a different (probably bigger,
higher-resolution) grid, so we have to use IceModelVec::regrid() here to
put this field onto the grid in use.